requirements.txt
streamlit
requests
pymupdf
python-docx
sentence-transformers
faiss-cpu
numpy
gdown
